WIP deprecate binary_codec_sv2 crate..
the only strict technical requirement for having a separate crate is derive_codec_sv2, which is dedicated to macro derivation

but as we are unifying the `binary_sv2` crate, there's no reason to keep a separate `binary_codec_sv2` and this commit is meant to merge its `codec` and `datatypes` modules into the root `binary_sv2` crate

this commit is still a WIP... I was able to get all the tests passing inside `binary_sv2` but things still break outside of it

@Shourya742 please complete this commit, amending to it when you are able to finish the migration
