+++ protocols/Cargo.toml
@@ -4,7 +4,6 @@ resolver="2"
 
 members = [
     "v1",
-    "v2/binary-sv2/codec",
     "v2/binary-sv2/derive_codec",
     "v2/binary-sv2",
     "v2/noise-sv2",
