--- protocols/v2/binary-sv2/codec/.gitignore
@@ -1,1 +1,0 @@
-/target
