--- protocols/v2/binary-sv2/codec/Cargo.toml
@@ -1,27 +1,0 @@
-[package]
-name = "binary_codec_sv2"
-version = "1.2.0"
-authors = ["The Stratum V2 Developers"]
-edition = "2018"
-readme = "README.md"
-description = "Sv2 data format"
-documentation = "https://docs.rs/binary_codec_sv2"
-license = "MIT OR Apache-2.0"
-repository = "https://github.com/stratum-mining/stratum"
-homepage = "https://stratumprotocol.org"
-keywords = ["stratum", "mining", "bitcoin", "protocol"]
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-quickcheck = { version = "1.0.0", optional = true }
-buffer_sv2 = { path = "../../../../utils/buffer", optional=true }
-
-[features]
-no_std = []
-default = ["no_std"]
-prop_test = ["quickcheck"]
-with_buffer_pool = ["buffer_sv2"]
-
-[package.metadata.docs.rs]
-features = ["with_buffer_pool"]
