+++ protocols/v2/binary-sv2/derive_codec/Cargo.toml
@@ -14,7 +14,7 @@ keywords = ["stratum", "mining", "bitcoin", "protocol"]
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-binary_codec_sv2 = { path="../codec"}
+#binary_codec_sv2 = { path="../codec"}
 
 [lib]
 proc-macro = true
