+++ protocols/v2/binary-sv2/src/codec/encodable.rs
