+++ protocols/v2/binary-sv2/src/datatypes/copy_data_types.rs
