+++ protocols/v2/binary-sv2/src/datatypes/non_copy_data_types/inner.rs
